js-ctypes-based Word for Windows integration
The C API here is loosely based on Word for Mac integration, while the code is very similar to our old Word for Windows integration. I updated the project to MSVC 2010, which produces a much larger dll, but that matters less now that there's only one for all Firefox versions.