Fix "Could not find a running Word instance" under certain circumstances
As noted at http://forums.zotero.org/discussion/25181/, Word gives us the
mapped path to a file on a mapped network drive, whereas it registers the
UNC path in the Running Object Table. Usually, this shouldn't matter, since
we fall back to looking for the Word application instance in the ROT if we
can't find the document, but if there are multiple running Word instances
only the first is registered in the ROT, so it could cause intermittent
issues for people whose documents are on mapped network drives.