New icons
Four changes:

1. Replaces icons with ones that resize (for user customization of
ribbon and Quick Access Toolbar) and invert colors (in Word 2016) more
aesthetically nicely. For remove field codes, I used a stock icon. For
Citation and Bibliography, I made icons that resemble the built-in Word
citation and bibliography buttons with Zotero branding.

2. Also includes secondary versions of the Citation, Bibliography, and
Doc Preferences macros with alternative sized icons (small or large).
These are useful for user customizations. It’s not possible to supply
both formats to the same ribbon button using Ribbon XML. There are some
paid add-in development packages that have created this sort of
functionality, but I’m not sure how they work.

3. Add Zotero app icon to the Ribbon group to show during automatic
resizing or when the group is added to QAT.

4. Combine the two bibliography functions into a single Add/Edit
Bibliography button. This would seem to simplify the user experience a
lot ala the combined citation button. This will require adding the
corresponding command to integration.js in zotero/zotero. I’m not
familiar enough with the integration.js to make that addition. Is there
interest in moving in this direction?

(I added the Add/Edit Bibliography command to Zotero.dot, but I don’t
have access to Word 2003 to edit the toolbar.)